refactor(core): reorganize test/ folder to mirror src/ structure
- Move test/data/ → test/helpers/ with builders/ and generators/ subdirs
- Rename NameData.ts → TextData.ts (class was already named TextData)
- Organize test/shared/ flat files into base/, errors/, i18n/, vo/ subdirs
- Fix remaining stale imports in Experience.test.ts and ProfileBuilder.ts
- Update all imports across test files and builder helpers
