+++ packages/core/test/helpers/builders/ExperienceBuilder.ts
@@ -5,8 +5,8 @@ import {
   EmploymentTypeValue,
   ILocalizedTextInput,
   LocationTypeValue,
-} from '../../src';
-import { Data } from './bases';
+} from '../../../src';
+import { Data } from '../generators';
 import { EntityBuilder } from './EntityBuilder';
 import { SkillBuilder } from './SkillBuilder';
