+++ packages/core/test/helpers/builders/index.ts
@@ -1,3 +1,4 @@
+export * from './EntityBuilder';
 export * from './ExperienceBuilder';
 export * from './LanguageBuilder';
 export * from './ProfileBuilder';
@@ -5,4 +6,3 @@ export * from './ProfessionalValueBuilder';
 export * from './ProjectBuilder';
 export * from './SkillBuilder';
 export * from './SocialNetworkBuilder';
-export * from './bases';
