+++ packages/core/test/helpers/generators/LocationTypeData.ts
