+++ packages/core/test/helpers/generators/SkillTypeData.ts
