+++ packages/core/test/helpers/generators/index.ts
@@ -2,9 +2,17 @@ import { EmploymentTypeData } from './EmploymentTypeData';
 import { FluencyData } from './FluencyData';
 import { ImageData } from './ImageData';
 import { LocationTypeData } from './LocationTypeData';
-import { TextData } from './NameData';
 import { SkillTypeData } from './SkillTypeData';
 import { SlugData } from './SlugData';
+import { TextData } from './TextData';
+
+export { EmploymentTypeData };
+export { FluencyData };
+export { ImageData };
+export { LocationTypeData };
+export { SkillTypeData };
+export { SlugData };
+export { TextData };
 
 export class Data {
   static employment = EmploymentTypeData;
