+++ packages/core/test/shared/vo/Url.test.ts
@@ -1,4 +1,4 @@
-import { Url, ValidationError } from '../../src';
+import { Url, ValidationError } from '../../../src';
 
 describe('Url', () => {
   describe('when created from valid value', () => {
